+++ lib/src/palette_generator.dart
@@ -2,7 +2,7 @@
 ///
 /// Use of this source code is governed by a MIT license that can be
 /// found in the LICENSE file.
-/// 
+///
 /// --------------------------------------------------------------------------------
 /// Copyright 2014 The Chromium Authors. All rights reserved.