+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Can generate Color palette to use in user interface and theming from wallpapers.
 - Can generate color palettes from an application icon.
 - Provided [Applications] to better use information from appinfo.
-- Can determine brightness of wallpaper (or any other image) using dominant colors and also has options to show brightness percentage
+- Can determine brightness of wallpaper (or any other image) using dominant colors.
 
 ## 0.1.1